update command system api
This commit renames the `SlashCommand` trait to `Command`, adds the
`command_type` required trait method to the trait as well as renaming
the method `execute_slash_command` to `execute`. This is reflected by
the changes to the existing implementations of the `Command` trait by
individual commands of the bot.

This commit changes the mechanism of how application commands are created
and registered. Now, instead of assuming that the command is a `CHAT_INPUT`
application command, the code now checks for the type of the command (whether
it is CHAT_INPUT, MESSAGE or USER, as described in the new `CommandType`
enumeration), then executes the corresponding method on the instance of
`CreateGlobalCommand` used to create the global application command.