Fix Directory Deletion
This commit fixes the deletion of the "dist" directory in "web-frontend"
when the buildsystem "clean" command is executed for the "web-frontend"
workspace.

Previously, executing "rmdir" from Process does not work because the
"rmdir" program is not found (on Windows I believe, because they are not
files). This commit fixes it by utilizing the filesystem API instead,
using the ".deleteRecursively()" function.